+++ 082/Problem082.cc
@@ -102,14 +102,15 @@ int main( int argc, char** argv )
 {
     vector<int> t_mat;
     int* i = test_mat;
-    while( *i >= 0 )
-    {
+    while( *i >= 0 ) {
         t_mat.push_back( *i );
         ++i;
     }
 
+/*
     cout << endl << "Test matrix: min path sum = "
          << min_path_sum( t_mat, 5 ) << endl;
+*/
 
     char* filename = (char*) "matrix.txt";
     if ( argc > 1 ) filename = argv[1];
@@ -118,7 +119,7 @@ int main( int argc, char** argv )
     parseMatrix( filename, mat );
     cout << "Size of mat: " << mat.size() << endl;
     int r = 1;
-    while( r*r < mat.size() ) ++r;
+    while( r*r < (int) mat.size() ) ++r;
     cout << "Root = " << r << endl;
     cout << endl << "Min path sum = " << min_path_sum( mat, r ) << endl << endl;
 }
